MobileData: Don't let wait-for-radio trample wait-for-data
Both waiting for radio and waiting for data used "is the progress
bar visible" as the "allow to continue" clause. If radio service
state changes arrived while the wait-for-data was ongoing, this'd
terminate the progress bar and make it wait forever.

Additionally, if wifi is connected, don't wait for data at all

Ref CYNGNOS-3126

Change-Id: I81792ac8be00c6a6746af88a62220c727cddc6d8